+++ tests/testthat-helpers.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(knmitransformer)
+
+test_check("knmitransformer", filter = "readInput")