+++ tests/testthat/test-evaporation.R
@@ -5,14 +5,14 @@ flog.threshold(DEBUG)
 flog.appender(appender.file('knmitransformer_evaporation.log'))
 library(data.table)
 
-ifile_tg   <- "regressionInput/KNMI14____ref_tg___19810101-20101231_v3.2_260.txt"
-ifile_rsds <- "regressionInput/KNMI14____ref_rsds___19810101-20101231_wrong2.txt"
+input_tg   <- "regressionInput/KNMI14____ref_tg___19810101-20101231_v3.2_260.txt"
+input_rsds <- "regressionInput/KNMI14____ref_rsds___19810101-20101231_wrong2.txt"
 ofile      <- NA
 regio.file <- "stationstabel" # table that links stations to region
 rounding   <- FALSE
 
 test_that("differences in header are discovered", {
-  expect_error(TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  expect_error(TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                 scenario = "GL",
                 horizon = 2030,
                 regio.file = regio.file,
@@ -22,10 +22,10 @@ test_that("differences in header are discovered", {
 
 context("evmk transformation - Entire station set")
 
-ifile_tg   <- system.file("refdata",
+input_tg   <- system.file("refdata",
                           "KNMI14____ref_tg___19810101-20101231_v3.2.txt",
                           package="knmitransformer")
-ifile_rsds <- system.file("refdata",
+input_rsds <- system.file("refdata",
                           "KNMI14____ref_rsds___19810101-20101231_v3.2.txt",
                           package="knmitransformer")
 ofile      <- NA
@@ -38,7 +38,7 @@ test_that("2030 decadal prediction", {
   horizon = 2030
 
 
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile="tmp.txt",
                        scenario=scenario,
                        horizon = horizon,
@@ -48,7 +48,7 @@ test_that("2030 decadal prediction", {
 
   expect_equal_to_reference(tmp, "regressionOutput/evaporation/KNMI14___2030_evmk.rds")
 
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile=ofile,
                        scenario=scenario,
                        horizon = horizon,
@@ -62,7 +62,7 @@ test_that("Scenario WL", {
   scenario="WL"
 
   horizon = 2050
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile=ofile,
                        scenario=scenario,
                        horizon = horizon,
@@ -72,7 +72,7 @@ test_that("Scenario WL", {
 
   # Difference of 0.07 instead of the usual 0.01
   horizon =  2085
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile=ofile,
                        scenario=scenario,
                        horizon = horizon,
@@ -85,7 +85,7 @@ test_that("Scenario WH", {
   scenario="WH"
 
   horizon = 2050
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile=ofile,
                        scenario=scenario,
                        horizon = horizon,
@@ -94,7 +94,7 @@ test_that("Scenario WH", {
   expect_equal_to_reference(tmp, "regressionOutput/evaporation/KNMI14_WH_2050_evmk.rds")
 
   horizon =  2085
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile=ofile,
                        scenario=scenario,
                        horizon = horizon,
@@ -107,7 +107,7 @@ test_that("Scenario GH", {
   scenario="GH"
 
   horizon = 2050
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile=ofile,
                        scenario=scenario,
                        horizon = horizon,
@@ -116,7 +116,7 @@ test_that("Scenario GH", {
   expect_equal_to_reference(tmp, "regressionOutput/evaporation/KNMI14_GH_2050_evmk.rds")
 
   horizon =  2085
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile=ofile,
                        scenario=scenario,
                        horizon = horizon,
@@ -129,7 +129,7 @@ test_that("Scenario GL", {
   scenario="GL"
 
   horizon = 2050
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile=ofile,
                        scenario=scenario,
                        horizon = horizon,
@@ -138,7 +138,7 @@ test_that("Scenario GL", {
   expect_equal_to_reference(tmp, "regressionOutput/evaporation/KNMI14_GL_2050_evmk.rds")
 
   horizon =  2085
-  tmp <- TransformEvap(ifile_tg=ifile_tg, ifile_rsds=ifile_rsds,
+  tmp <- TransformEvap(input_tg=input_tg, input_rsds=input_rsds,
                        ofile=ofile,
                        scenario=scenario,
                        horizon = horizon,